# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: HELBOOT API
  description: >
    REST API of the HELBOOT provisioning engine. The web UI communicates
    exclusively through this API; it is equally intended for automation.
    Error responses always use the shared Error envelope whose `code` is a
    stable identifier doubling as the i18n key.
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
...
paths:
...
  /system/info:
    get:
      tags: [system]
      summary: System information
      responses:
        "200":
          description: Version, setup state and network mode
          content:
            application/json:
              schema:
                type: object
                properties:
                  version: { type: string }
                  setupCompleted: { type: boolean }
                  networkMode: { type: string, enum: ["", proxy_dhcp, dhcp] }
                  providers: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /setup:
    post:
      tags: [setup]
      summary: Complete the first-run wizard
      description: Unauthenticated by design, usable exactly once.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [language, admin, network]
              properties:
                language: { type: string, enum: [en, de] }
                admin:
                  type: object
                  required: [username, password]
                  properties:
                    username: { type: string, minLength: 3 }
                    password: { type: string, minLength: 10 }
                network:
                  type: object
                  required: [mode]
                  properties:
                    mode: { type: string, enum: [proxy_dhcp, dhcp] }
      responses:
        "201": { description: Setup completed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Setup was already completed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with local credentials
      description: Rate-limited per client IP. Sets the session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { description: Rate limited, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /auth/logout:
    post:
      tags: [auth]
      summary: Log out and destroy the session
      responses:
        "200": { description: Logged out }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/me:
    get:
      tags: [auth]
      summary: Current user and CSRF token
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/password:
    post:
      tags: [auth]
      summary: Change your own password
      description: Verifies the current password; all sessions are revoked on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 10 }
      responses:
        "200": { description: Changed; re-login required }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { description: Current password incorrect, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /users:
    get:
      tags: [users]
      summary: List users (admin)
      responses:
        "200":
          description: All accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403": { description: Requires admin role, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      tags: [users]
      summary: Create a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string, minLength: 3 }
                password: { type: string, minLength: 10 }
                role: { type: string, enum: [admin, operator, viewer] }
                locale: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Username taken, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    put:
      tags: [users]
      summary: Change a user's role (admin)
      description: The last administrator cannot be demoted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [admin, operator, viewer] }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Last administrator, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    delete:
      tags: [users]
      summary: Delete a user (admin)
      description: Self-deletion and removing the last administrator are rejected.
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Self or last administrator, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /users/{id}/password:
    put:
      tags: [users]
      summary: Reset a user's password (admin)
      description: All sessions of the user are revoked.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 10 }
      responses:
        "200": { description: Updated }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /providers:
    get:
      tags: [providers]
      summary: List provider manifests
      description: >
        The UI derives all OS-specific options from these capability
        declarations; nothing is hardcoded per OS.
      responses:
        "200":
          description: All loaded providers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Provider" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /providers/{name}:
    get:
      tags: [providers]
      summary: Get one provider manifest
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The provider manifest
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "404": { $ref: "#/components/responses/NotFound" }
  /hosts:
    get:
      tags: [hosts]
      summary: List hosts
      responses:
        "200":
          description: All hosts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Host" }
    post:
      tags: [hosts]
      summary: Register a host (operator)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HostRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Host" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: MAC already registered, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /hosts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [hosts]
      summary: Get a host
      responses:
        "200":
          description: The host
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Host" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [hosts]
      summary: Update a host (operator)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HostRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Host" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [hosts]
      summary: Delete a host (operator)
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /profiles:
    get:
      tags: [profiles]
      summary: List profiles
      responses:
        "200":
          description: All profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Profile" }
    post:
      tags: [profiles]
      summary: Create a profile (operator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, provider]
              properties:
                name: { type: string }
                provider: { type: string }
                isoId: { type: [integer, "null"] }
                config: { type: object, description: Provider-shaped configuration document }
      responses:
        "201":
          description: Created with version 1
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /profiles/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [profiles]
      summary: Get a profile
      responses:
        "200":
          description: The profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [profiles]
      summary: Update a profile (operator)
      description: >
        Sending `config` edits the current version in place; with
        `saveAsNewVersion` it becomes an immutable new version
        (ADR-0013). In-place edits of a version referenced by an
        installation are refused with 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                isoId: { type: [integer, "null"] }
                config: { type: object }
                saveAsNewVersion: { type: boolean, default: false }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Version is referenced by an installation (`profile.version_in_use`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [profiles]
      summary: Delete a profile (operator)
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /profiles/{id}/versions:
    get:
      tags: [profiles]
      summary: List a profile's version snapshots
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: All versions, oldest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProfileVersion" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /profiles/{id}/versions/{version}/answer:
    get:
      tags: [profiles]
      summary: Preview the rendered answer file of a profile version
      description: >
        Renders the answer file exactly as the boot chain would, with
        recognizable placeholder boot parameters (ADR-0014). Returns the
        manual override when one is set.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: version, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: The rendered answer file
          content:
            application/json:
              schema:
                type: object
                properties:
                  format: { type: string }
                  content: { type: string }
                  overridden: { type: boolean }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: The template (or override) fails to render (`profile.answer_render_failed`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /profiles/{id}/versions/{version}/answer-override:
    put:
      tags: [profiles]
      summary: Set or clear the manual answer-file override (operator)
      description: >
        A non-empty `content` replaces the provider template for this
        version until cleared; it runs through the same template engine,
        so boot parameters like `{{ .ISOURL }}` keep working. An empty
        `content` clears the override (ADR-0014).
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: version, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        "204": { description: Stored }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: The override fails to render (`profile.answer_render_failed`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /logs:
    get:
      tags: [logs]
      summary: Recent log entries (newest first)
      parameters:
        - name: level
          in: query
          schema: { type: string, enum: [debug, info, warn, error], default: info }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        "200":
          description: Buffered log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    time: { type: string, format: date-time }
                    level: { type: string }
                    message: { type: string }
                    attrs:
                      type: object
                      additionalProperties: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /backup/export:
    get:
      tags: [backup]
      summary: Download a full-state backup archive (admin)
      description: >
        tar.gz with a consistent database snapshot (settings, users,
        hosts, profiles with versions, installation history, ISO
        checksums). ISO files are excluded by size; their checksums allow
        detecting missing images after a restore.
      responses:
        "200":
          description: Backup archive
          content:
            application/gzip:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /backup/import:
    post:
      tags: [backup]
      summary: Stage a backup archive for restore (admin)
      description: The restore is applied on the next restart.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Staged
          content:
            application/json:
              schema:
                type: object
                properties:
                  staged: { type: boolean }
                  restartRequired: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
  /installations:
    get:
      tags: [installations]
      summary: List the installation queue and history
      responses:
        "200":
          description: All installations, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Installation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [installations]
      summary: Queue an installation (operator)
      description: >
        The next network boot of the host starts the installer. Pins the
        profile's current version snapshot so later profile edits never
        change what a queued host installs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostId]
              properties:
                hostId: { type: integer }
                profileId:
                  type: [integer, "null"]
                  description: Overrides the host's assigned profile.
      responses:
        "201":
          description: Queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Host already has an active installation, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /installations/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [installations]
      summary: Get an installation
      responses:
        "200":
          description: The installation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [installations]
      summary: Cancel a waiting installation (operator)
      responses:
        "204": { description: Cancelled }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Not cancellable (already running or finished), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /isos:
    get:
      tags: [isos]
      summary: List ISO images
      responses:
        "200":
          description: All ISO records with analysis results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ISOImage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /isos/upload:
    post:
      tags: [isos]
      summary: Upload an original ISO (operator)
      description: >
        Streaming multipart upload (single part named "file"). The image
        is hashed, stored unmodified and analyzed against the provider
        detection rules.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Stored and analyzed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ISOImage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: Filename already exists, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /isos/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [isos]
      summary: Get an ISO record
      responses:
        "200":
          description: The ISO record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ISOImage" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [isos]
      summary: Delete an ISO record and its file (operator)
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /network/config:
    get:
      tags: [network]
      summary: Boot-network configuration
      responses:
        "200":
          description: Current configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NetworkConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [network]
      summary: Update the boot-network configuration (admin)
      description: Changes take effect after a restart.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NetworkConfig" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved: { type: boolean }
                  restartRequired: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: helboot_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on all mutating requests.
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable identifier, used as the frontend i18n key.
            message:
              type: string
              description: English fallback message.
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role: { type: string, enum: [admin, operator, viewer] }
        locale: { type: string }
        createdAt: { type: string, format: date-time }
    SessionInfo:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        csrfToken: { type: string }
    Provider:
      type: object
      properties:
        name: { type: string }
        displayName: { type: string }
        family: { type: string }
        capabilities:
          type: object
          additionalProperties: { type: boolean }
        answerFile:
          type: object
          properties:
            format: { type: string }
            template: { type: string }
        settingsSchema:
          type: array
          description: Field descriptors the UI renders the profile form from (ADR-0012).
          items: { $ref: "#/components/schemas/SettingsField" }
        detection:
          type: object
          properties:
            volumeIdPatterns: { type: array, items: { type: string } }
            files: { type: array, items: { type: string } }
        boot:
          type: object
          additionalProperties:
            type: object
            properties:
              kernel: { type: string }
              initrd: { type: array, items: { type: string } }
              cmdline: { type: string }
              requires: { type: array, items: { type: string } }
        notes: { type: string }
...
    Host:
      allOf:
        - $ref: "#/components/schemas/HostRequest"
        - type: object
          properties:
            id: { type: integer }
            status: { type: string, enum: [discovered, ready, installing, error] }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    Profile:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        provider: { type: string }
        isoId: { type: [integer, "null"] }
        currentVersion: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
security:
  - sessionCookie: []